pyproject.toml
[project]
name = "epaper-target"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "opencv-python",
    "numpy",
    "pygame",
    "pyyaml",
]
